deal with the unfortunate fact that family names contain a lower-case ß (German sharp s)
This seems a necessary fix to recognise family names containing an ß, which could be capitalised as SS – some people use SZ, or since 2008 [Unicode] or 2017 [official German rules] there is ẞ – but the EP seems to think it best to use a lower-case ß all the same. False positives would arise if there were German first names consisting of just an initial (hence upper-case) letter plus an ß. But that does not seem be the case (says a native German).